pyproject.toml
[project]
name = "agentes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "faster-whisper>=1.2.1",
    "flask>=3.1.3",
    "kokoro-onnx>=0.5.0",
    "langchain>=1.2.15",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.14",
    "langgraph>=1.1.8",
    "mcp>=1.27.0",
    "numpy>=2.4.4",
    "ollama>=0.6.1",
    "onnxruntime-directml>=1.24.4",
    "openai>=2.32.0",
    "psutil>=7.2.2",
    "pyaudio>=0.2.14",
    "pyfiglet>=1.0.4",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scipy>=1.17.1",
    "sounddevice>=0.5.5",
]
